Fix scrolling bugs on lists
The search hint resting place was pushing it outside the right hand edge
of the screen causing a horizontal scrollbar to appear when it was not
required. This change aligns it the right hand edge so it doesn't add a
horizontal scroll bar or interfere with the search input.

The additional `<span />` element used to measure the width of the text
was causing a vertical scroll bar to get added so that's now set to
`display: none;`. The measurement still works just fine in Firefox,
Chrome and Safari but I've not had chance to test it in IE or Edge.