don't clear existing date values
Problem:

When a user edits a resource
with a `Date`, `Time`, or `DateTime` attribute,
and the attribute already has a value,
the date-time picker clears out the value.

Unless the user re-fills the field,
the value will be reset and will cause an error
when they hit "submit".

Solution:

Update the `datetimepicker` date format
to match the Rails default.

This solution introduced a secondary problem:

With the datetimepicker format change,
tests fail intermittently in `spec/features/orders_form_spec.rb`.

It looks like somehow, DatabaseCleaner is running early
after the `update` action redirects or renders a new view.
It is difficult to track down the exact failure, however.

Solution:

Switch to poltergeist as a js test runner.
Not sure why this stops the tests from failing, but it does.