Select correct dashboards for polymorphic fields
Fixes #245

Problem:

Polymorphic relationships try to look up a Dashboard
for the associated object
based on the relationship's attribute name.

In most cases, the associated object's class (and therefore dashboard)
will be diferent from the relationship name,
so the lookup will fail.

Solution:

For polymorphic relationships,
derive the dashboard name
from the class of the associated object (if it exists).

Minor changes:

This also updates the default polymorphic views to use the configurable
namespace.