Use `Administrate::Namespace` to detect resources
Problem:

Administrate's generated routes have a dependency on the
`DashboardManifest`. Often, the generated routes are loaded without
`DashboardManfiest` being defined, which throws an error.

Solution:

Remove the `DashboardManifest` dependency from the generated routes.

In order to do this, we will explicitly define the routes in
`config/routes.rb`, and remove the `DashboardManifest` entirely.

This approach better fits Rails conventions
and narrows the public API of Administrate.