Merge the existing translations before testing. (#772)
`with_translations` is used to test that a view does actually implement
a given translation. Previously, we were resetting the known
translations to only use the one under test. We now have many more
translations so this approach will fail when visiting a page because of
"missing translations".

Example: https://git.io/vywpc

This came up when merging #344.