Fix `aria-labelledby` on tables (#886)
`aria-labelledby` accepts `id`s of other elements which act as an
accessible label(s). It's great for providing a label to tables, which
are large amounts of content; it helps people understand what the table
content actually is.

We're currently using `aria-labelledby` on tables in Administrate,
pointing to the `id` of the `h1` on index pages. However, this breaks
when tables are shown as attribute data on show pages, because the `id`
of `page-title` does not exist, and it also is not the proper label for
the table (the attribute name is). You can see an example of this here:
https://administrate-prototype.herokuapp.com/admin/customers/9849

This PR allows an `id` of an element to be passed into the partial
which displays the collection table, so that it can used as the value
for the table's `aria-labelledby` attribute.