Add additional Rubies to CI using Workflows. (#1032)
A recent issue (#1010) highlighted a deprecation under certain Ruby
versions that we should be aware of. So far, we've been using CircleCI
for running tests and previously they didn't support multiple builds.

With Circle CI 2.0, Workflows allows for something like this to be setup. This
uses YAML inheritance to share a job definition across multiple build
types (defined as different Docker configurations).

This introduces Ruby 2.3 and 2.4.3 in addition to 2.5.0 and using Postgres 9.5.

Notable caveat to this approach is that you must set different databases
for each environment (here named like the Ruby versions) to avoid test
state-based errors.